#240: removed css from sidebar and navigation
