+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite --host",
     "copy-assets": "cpx src/assets/**/* dist/assets",
     "build": "vite build && tsc -p tsconfig.build.json && tsc-alias -p tsconfig.build.json && npm run copy-assets",
     "clean": "rm -rf dist",
