+++ src/webui/components/application/Footer.tsx
@@ -1,0 +1,13 @@
+
+import { Row } from '^/webui/designsystem';
+
+type Props = {
+
+};
+
+export default function Component({ }: Props)
+{
+    return <Row alignX='justify' alignY='center'>
+        Footer
+    </Row>;
+}
