--- src/webui/components/application/navigation/Item.css
@@ -1,4 +1,0 @@
-.navigation-item
-{
-    text-decoration: none;
-}
