--- src/webui/components/application/sidebar/Sidebar.css
@@ -1,43 +1,0 @@
-.application-sidebar
-{
-    display: flex;
-    flex-direction: row;
-    gap: 2em;
-    height: 100%;
-    width: 100%;
-
-    & > .content
-    {
-        display: flex;
-        flex-direction: column;
-        flex-grow: 1;
-        flex-shrink: 1;
-        gap: 2rem;
-        overflow: hidden;
-
-        .logo
-        {
-            flex-grow: 0;
-            flex-shrink: 0;
-        }
-
-        nav
-        {
-            flex-grow: 1;
-            flex-shrink: 1;
-            overflow-y: auto;
-        }
-
-        .identity
-        {
-            flex-grow: 0;
-            flex-shrink: 0;
-        }
-    }
-
-    & > .ruler
-    {
-        flex-grow: 0;
-        flex-shrink: 0;
-    }
-}
