+++ src/webui/components/index.ts
@@ -1,9 +1,10 @@
 
+export { default as ApplicationFooter } from './application/Footer';
 export { default as ApplicationHeader } from './application/Header';
 export { default as ApplicationIntroduction } from './application/Introduction';
 export { default as ApplicationLegalInfo } from './application/LegalInfo';
 export { default as ApplicationLogo } from './application/Logo';
-export { default as ApplicationSidebar } from './application/sidebar/Sidebar';
+export { default as ApplicationSidebar } from './application/Sidebar';
 export { default as ComicEditor } from './comic/Editor';
 export { default as CommentForm } from './comment/Form';
 export { default as ErrorBoundary } from './common/ErrorBoundary';
@@ -21,3 +22,4 @@ export { default as PostPanelList } from './post/PanelList';
 export { default as ReactionPanelList } from './reaction/PanelList';
 export { default as RelationPanelList } from './relation/PanelList';
 export { default as RelationProfile } from './relation/Profile';
+
