+++ src/webui/designsystem/elements/cell/Cell.css
@@ -1,0 +1,19 @@
+.ds
+{
+    .cell
+    {
+        &.sizing-fixed
+        {
+            flex-grow: 0;
+            flex-shrink: 0;
+            overflow: hidden;
+        }
+
+        &.sizing-fluid
+        {
+            flex-grow: 1;
+            flex-shrink: 1;
+            overflow: auto;
+        }
+    }
+}
