+++ src/webui/layouts/sidebar/Sidebar.css
@@ -19,7 +19,6 @@
     {
         display: flex;
         flex-direction: row;
-        gap: 2em;
         width: 100vw;
         height: 100vh;
         max-height: 100vh;
@@ -33,6 +32,7 @@
 
             aside
             {
+                display: flex;
                 flex-shrink: 0;
                 flex-grow: 0;
                 padding: 1em 0;
