+++ src/webui/pages/Application.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { ApplicationHeader, ApplicationSidebar } from '^/webui/components';
+import { ApplicationFooter, ApplicationHeader, ApplicationSidebar } from '^/webui/components';
 import { useAppContext } from '^/webui/contexts/AppContext';
 import { SidebarLayout } from '^/webui/layouts';
 
@@ -23,9 +23,10 @@ export default function Page({ children }: Props)
     const onLogout = () => navigate('/logout');
 
     const header = <ApplicationHeader identity={identity} onLogout={onLogout} />;
+    const footer = <ApplicationFooter />;
     const sidebar = <ApplicationSidebar identity={identity} onLogout={onLogout} />;
 
-    return <SidebarLayout header={header} sidebar={sidebar}>
+    return <SidebarLayout header={header} footer={footer} sidebar={sidebar}>
         {children}
     </SidebarLayout>;
 }
